Rename feature to `select-notifications`
